Unify F2Z arithmetic under vendor/field
Centralize prime, binary-field and fixed-width integer arithmetic for F2Z,
circuit, and local Flock in vendor/field, retaining optimized kernels and
using one production delayed reduction path.

Include the standalone Binius benchmark dependency fix and remove generated
validation logs, experiment results, and source archives.

Performance qualification remains pending.

+++ crates/circuit/src/sha256.rs
@@ -547,7 +547,6 @@ mod tests {
     use std::iter::Sum;
     use std::ops::{Add, AddAssign};
 
-    use num_bigint::BigInt;
     use num_traits::Zero;
 
     use super::*;
@@ -686,7 +685,7 @@ mod tests {
         for (row, value) in from_m.iter().enumerate() {
             assert_eq!(
                 value,
-                &BigInt::from(recorded.bit(row)),
+                &recorded.bit(row),
                 "integer witness differs at M row {row}"
             );
         }
